w32functions.c: fix compiler warning

+++ src/platform/w32functions.c
@@ -653,7 +653,7 @@ int W32_snprintf(char *__restrict s, size_t n, const char *__restrict format, ..
     va_start(args, format);
     ret = _vsnprintf(s, n, format, args);
     va_end(args);
-    if (n == ret)
+    if ((int)n == ret)
       s[n - 1] = 0;
     if (ret >= 0)
       return ret;
